+++ src/konpy/routes.clj
@@ -1,13 +1,16 @@
 (ns konpy.routes
   (:require [reitit.ring :as reitit-ring]
-            #_[ring.util.response :as response]
             [ring.middleware.defaults :refer [wrap-defaults site-defaults]]
             [taoensso.telemere :as t]
-            [konpy.assignments :as ka]
-            [konpy.example :as example]
+            ;
+            [konpy.tasks :as tasks]
+            [konpy.admin :as admin]
+            [konpy.answers :as answers]
             [konpy.login :refer [login-page login! logout!]]
-            [konpy.views :refer [under-construction]]
-            [konpy.middleware :as km]))
+            [konpy.utils :refer [under-construction-page]]
+            [konpy.middleware :as m]
+            ;
+            [konpy.example :as example]))
 
 ; FIXME: everytime compile.
 (defn routes
@@ -15,23 +18,26 @@
   [""
    ["/assets/*" (reitit-ring/create-resource-handler
                  {:path "/" :root "public"})]
-   ["/example" {:get {:handler example/example-page}
-                :post {:handler example/example-post}}]
    ["/" {:get  {:handler login-page}
          :post {:handler login!}}]
    ["/logout" logout!]
-   ["/assignments" {:middleware [km/wrap-users]}
-    ["/" ka/task]
-    #_["/tasks" ka/tasks]]
-   ["/answers" {:middleware [km/wrap-users]}
-    ["/" under-construction]]
-   ["/admin" {:middleware [km/wrap-admin]}
-    ["/" ka/tasks]
-    ["/edit" {:get ka/edit
-              :post ka/edit!}]
-    ["/delete" {:delete ka/delete!}]
-    ["/new" {:get ka/new
-             :post ka/create!}]]])
+   ["/tasks" {:middleware [m/wrap-users]}
+    ["/" tasks/tasks-this-week]
+    ["/all" tasks/tasks-all]]
+   ["/answers" {:middleware [m/wrap-users]}
+    ["/" under-construction-page]]
+   ["/admin" {:middleware [m/wrap-admin]}
+    ["/" {:get {:handler admin/tasks}}]
+    ["/new" {:get {:handler admin/new}
+             :post {:handler admin/create!}}]]
+   ["/edit/:n" {:get admin/edit
+                :post admin/edit!}]
+   ["/delete/:n" {:delete admin/delete!}]
+   ;
+   ["/example"
+    ["" {:get  {:handler example/example-page}
+         :post {:handler example/example-post}}]
+    ["/confirm" {:get {:handler example/example-confirm}}]]])
 
 (defn not-found-handler
   [_]