+++ src/konpy/utils.clj
@@ -1,0 +1,32 @@
+(ns konpy.utils
+  (:require
+   [java-time.api :as jt]
+   [konpy.views :refer [page]]))
+
+; views namespace?
+(defn under-construction-page [_]
+  (page
+   [:div {:class ""} "under construction"]))
+
+(defn now []
+  (str (jt/local-date)))
+
+(comment
+  (now)
+  (jt/instant)
+  (java.util.Date.))
+(defn user [request]
+  (get-in request [:session :identity]))
+
+(def start-day (jt/local-date 2025 3 31))
+
+(defn weeks
+  "Returns how many weeks have passed since the argument `date`.
+   If no argument given, use the `start`day` defnied above."
+  ([] (weeks (jt/local-date)))
+  ([date]
+   (quot (jt/time-between start-day date :days) 7)))
+
+(comment
+  (weeks (jt/local-date 2025 6 6))
+  :rcf)